+++ Source/PakDoc.cpp
@@ -215,7 +215,7 @@ void CPakDoc::PasteFiles()
 			}
 
 			// Determine the data size
-			int iDataSize = GlobalSize(hgData);
+			SIZE_T iDataSize = GlobalSize(hgData);
 						
 			// Alloc memory
 			BYTE *pbyData = (BYTE *)GlobalAlloc(GMEM_FIXED, iDataSize);
